requirements.txt
streamlit
numpy<2.0
sentence-transformers==2.2.2
transformers==4.30.2
torch==2.2.2
faiss-cpu==1.7.4
rank-bm25>=0.2.2
tika
spacy>=3.7
pytesseract
pdf2image
pillow
openai>=1.0.0
streamlit>=1.39
huggingface-hub<0.15
